added new model architectures
