+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project uses [Semantic Versioning](http://semver.org/).
 
+# [0.X.0] - 2021-5-12
+### Added
+ - new model architectures ``CollaborativeMetricLearningModel``, ``MLPMatrixFactorizationModel``, and ``DeepFM``
+### Changed
+ - filename for ``HybridPretrainedModel`` to ``hybrid_pretrained_matrix_factorization.py``. The former model filepath is now deprecated and will be removed in future version ``0.X.0``
+ - all ``model.freeze() -> model.eval()``
+
 # [0.3.0] - 2021-5-10
 ### Changed
 - Disable automated batching in ApproximateNegativeSamplingInteractionsDataLoader and  HDF5InteractionsDataLoader.
