+++ collie_recs/model/__init__.py
@@ -1,5 +1,8 @@
 from collie_recs.model.base import *
+from collie_recs.model.collaborative_metric_learning import *
+from collie_recs.model.deep_fm import *
+from collie_recs.model.hybrid_pretrained_matrix_factorization import *
 from collie_recs.model.matrix_factorization import *
-from collie_recs.model.nonlinear_matrix_factorization import *
+from collie_recs.model.mlp_matrix_factorization import *
 from collie_recs.model.neural_collaborative_filtering import *
-from collie_recs.model.hybrid_matrix_factorization_model import *
+from collie_recs.model.nonlinear_matrix_factorization import *
