+++ collie_recs/model/collaborative_metric_learning.py
@@ -3,43 +3,34 @@ from typing import Callable, Dict, Optional, Tuple, Union
 
 import numpy as np
 import torch
-from torch import nn
+import torch.nn.functional as F
 from torch.optim.lr_scheduler import ReduceLROnPlateau
 
 from collie_recs.model.base import (BasePipeline,
                                     interactions_like_input,
-                                    ScaledEmbedding,
-                                    ZeroEmbedding)
+                                    ScaledEmbedding)
 from collie_recs.utils import get_init_arguments
 
 
-class MatrixFactorizationModel(BasePipeline):
+class CollaborativeMetricLearningModel(BasePipeline):
     """
-    Training pipeline for the matrix factorization model.
+    Training pipeline for the collaborative metric learning model.
 
-    ``MatrixFactorizationModel`` models have an embedding layer for both users and items which are
-    dot-producted together to output a single float ranking value.
+    ``CollaborativeMetricLearningModel`` models have an embedding layer for both users and items. A
+    single float, prediction is retrieved by taking the pairwise distance between the two
+    embeddings.
 
-    Collie adds a twist on to this incredibly popular framework by allowing separate optimizers
-    for embeddings and bias terms. With larger datasets and multiple epochs of training, a model
-    might incorrectly learn to only optimize the bias terms for a quicker path towards a local
-    loss minimum, essentially memorizing how popular each item is. By using a separate, slower
-    optimizer for the bias terms (like Stochastic Gradient Descent), the model must prioritize
-    optimizing the embeddings for meaningful, more varied recommendations, leading to a model
-    that is able to achieve a much lower loss. See the documentation below for ``bias_lr`` and
-    ``bias_optimizer`` input arguments for implementation details.
-
-    All ``MatrixFactorizationModel`` instances are subclasses of the ``LightningModule`` class
-    provided by PyTorch Lightning. This means to train a model, you will need a
+    All ``CollaborativeMetricLearningModel`` instances are subclasses of the ``LightningModule``
+    class provided by PyTorch Lightning. This means to train a model, you will need a
     ``collie_recs.model.CollieTrainer`` object, but the model can be saved and loaded without this
     ``Trainer`` instance. Example usage may look like:
+    may look like:
 
     .. code-block:: python
 
-        from collie_recs.model import CollieTrainer, MatrixFactorizationModel
+        from collie.model import CollaborativeMetricLearningModel, CollieTrainer
 
-
-        model = MatrixFactorizationModel(train=train)
+        model = CollaborativeMetricLearningModel(train=train)
         trainer = CollieTrainer(model)
         trainer.fit(model)
         model.eval()
@@ -47,7 +38,7 @@ class MatrixFactorizationModel(BasePipeline):
         # do evaluation as normal with ``model``
 
         model.save_model(filename='model.pth')
-        new_model = MatrixFactorizationModel(load_model_path='model.pth')
+        new_model = CollaborativeMetricLearningModel(load_model_path='model.pth')
 
         # do evaluation as normal with ``new_model``
 
@@ -63,15 +54,11 @@ class MatrixFactorizationModel(BasePipeline):
         that when the model class is saved, datasets will NOT be saved as well
     embedding_dim: int
         Number of latent factors to use for user and item embeddings
-    dropout_p: float
-        Probability of dropout
     sparse: bool
         Whether or not to treat embeddings as sparse tensors. If ``True``, cannot use weight decay
         on the optimizer
     lr: float
         Embedding layer learning rate
-    bias_lr: float
-        Bias terms learning rate. If 'infer', will set equal to ``lr``
     lr_scheduler_func: torch.optim.lr_scheduler
         Learning rate scheduler to use during fitting
     weight_decay: float
@@ -84,12 +71,6 @@ class MatrixFactorizationModel(BasePipeline):
         * ``'adam'`` (for ``torch.optim.Adam``)
 
         * ``'sparse_adam'`` (for ``torch.optim.SparseAdam``)
-
-    bias_optimizer: torch.optim or str
-        Optimizer for the bias terms. This supports the same string options as ``optimizer``, with
-        the addition of ``infer``, which will set the optimizer equal to ``optimizer``. If
-        ``bias_optimizer`` is ``None``, only a single optimizer will be created for all model
-        parameters
     loss: function or str
         If a string, one of the following implemented losses:
 
@@ -136,16 +117,13 @@ class MatrixFactorizationModel(BasePipeline):
                  train: interactions_like_input = None,
                  val: interactions_like_input = None,
                  embedding_dim: int = 30,
-                 dropout_p: float = 0.0,
                  sparse: bool = False,
                  lr: float = 1e-3,
-                 bias_lr: Optional[Union[float, str]] = 1e-2,
                  lr_scheduler_func: Optional[Callable] = partial(ReduceLROnPlateau,
                                                                  patience=1,
                                                                  verbose=True),
                  weight_decay: float = 0.0,
                  optimizer: Union[str, Callable] = 'adam',
-                 bias_optimizer: Optional[Union[str, Callable]] = 'sgd',
                  loss: Union[str, Callable] = 'hinge',
                  metadata_for_loss: Optional[Dict[str, torch.tensor]] = None,
                  metadata_for_loss_weights: Optional[Dict[str, float]] = None,
@@ -158,22 +136,15 @@ class MatrixFactorizationModel(BasePipeline):
         """
         Method for building model internals that rely on the data passed in.
 
-        This method will be called after ``prepare_data``.
+        This method will be called after `prepare_data`.
 
         """
-        self.user_biases = ZeroEmbedding(num_embeddings=self.hparams.num_users,
-                                         embedding_dim=1,
-                                         sparse=self.hparams.sparse)
-        self.item_biases = ZeroEmbedding(num_embeddings=self.hparams.num_items,
-                                         embedding_dim=1,
-                                         sparse=self.hparams.sparse)
         self.user_embeddings = ScaledEmbedding(num_embeddings=self.hparams.num_users,
                                                embedding_dim=self.hparams.embedding_dim,
                                                sparse=self.hparams.sparse)
         self.item_embeddings = ScaledEmbedding(num_embeddings=self.hparams.num_items,
                                                embedding_dim=self.hparams.embedding_dim,
                                                sparse=self.hparams.sparse)
-        self.dropout = nn.Dropout(p=self.hparams.dropout_p)
 
     def forward(self, users: torch.tensor, items: torch.tensor) -> torch.tensor:
         """
@@ -181,9 +152,9 @@ class MatrixFactorizationModel(BasePipeline):
 
         Simple matrix factorization for a single user and item looks like:
 
-        ````prediction = (user_embedding * item_embedding) + user_bias + item_bias````
+        ```prediction = (user_embedding * item_embedding) + user_bias + item_bias```
 
-        If dropout is added, it is applied to the two embeddings and not the biases.
+        If dropout is added, it is applied for the two embeddings and not the biases.
 
         Parameters
         ----------
@@ -201,18 +172,7 @@ class MatrixFactorizationModel(BasePipeline):
         user_embeddings = self.user_embeddings(users)
         item_embeddings = self.item_embeddings(items)
 
-        preds = (
-            torch.mul(self.dropout(user_embeddings), self.dropout(item_embeddings)).sum(axis=1)
-            + self.user_biases(users).squeeze(1)
-            + self.item_biases(items).squeeze(1)
-        )
-
-        if self.hparams.y_range is not None:
-            preds = (
-                torch.sigmoid(preds)
-                * (self.hparams.y_range[1] - self.hparams.y_range[0])
-                + self.hparams.y_range[0]
-            )
+        preds = F.pairwise_distance(user_embeddings, item_embeddings)
 
         return preds
