+++ tests/fixtures/model_fixtures.py
@@ -7,9 +7,12 @@ import torch
 from collie_recs.interactions import (ApproximateNegativeSamplingInteractionsDataLoader,
                                       HDF5InteractionsDataLoader,
                                       InteractionsDataLoader)
-from collie_recs.model import (CollieTrainer,
+from collie_recs.model import (CollaborativeMetricLearningModel,
+                               CollieTrainer,
+                               DeepFM,
                                HybridPretrainedModel,
                                MatrixFactorizationModel,
+                               MLPMatrixFactorizationModel,
                                NeuralCollaborativeFiltering,
                                NonlinearMatrixFactorizationModel)
 from collie_recs.utils import pandas_df_to_hdf5
@@ -29,7 +32,7 @@ def implicit_model(train_val_implicit_data, gpu_count):
                                   logger=False,
                                   checkpoint_callback=False)
     model_trainer.fit(model)
-    model.freeze()
+    model.eval()
 
     return model
 
@@ -52,6 +55,10 @@ def untrained_implicit_model_no_val_data(train_val_implicit_data):
 
 @pytest.fixture(params=['mf_hdf5',
                         'mf_with_y_range',
+                        'collaborative_metric_learning',
+                        'sparse_collaborative_metric_learning',
+                        'mlp_mf',
+                        'mlp_mf_with_y_range',
                         'sparse_mf',
                         'mf_no_val',
                         'mf_non_approximate',
@@ -63,6 +70,11 @@ def untrained_implicit_model_no_val_data(train_val_implicit_data):
                         'neucf_relu',
                         'neucf_leaky_rulu',
                         'neucf_custom',
+                        'deep_fm',
+                        'deep_fm_sigmoid',
+                        'deep_fm_relu',
+                        'deep_fm_leaky_rulu',
+                        'deep_fm_custom',
                         'hybrid_pretrained',
                         'hybrid_pretrained_metadata_layers'])
 def models_trained_for_one_step(request,
@@ -120,7 +132,7 @@ def models_trained_for_one_step(request,
                                           checkpoint_callback=False)
 
             model_trainer.fit(model)
-            model.freeze()
+            model.eval()
 
             return model
 
@@ -165,6 +177,37 @@ def models_trained_for_one_step(request,
         model = MatrixFactorizationModel(train=train,
                                          val=val,
                                          y_range=(0, 4))
+    elif request.param == 'collaborative_metric_learning':
+        model = CollaborativeMetricLearningModel(train=train,
+                                                 val=val,
+                                                 embedding_dim=15,
+                                                 lr=1e-1,
+                                                 optimizer='adam',
+                                                 loss='hinge')
+    elif request.param == 'sparse_collaborative_metric_learning':
+        model = CollaborativeMetricLearningModel(train=train,
+                                                 val=val,
+                                                 embedding_dim=15,
+                                                 lr=1e-1,
+                                                 optimizer='sparse_adam',
+                                                 loss='hinge',
+                                                 sparse=True)
+    elif request.param == 'mlp_mf':
+        model = MLPMatrixFactorizationModel(train=train,
+                                            val=val,
+                                            embedding_dim=15,
+                                            num_layers=3,
+                                            dropout_p=0.1,
+                                            lr=1e-1,
+                                            bias_lr=1e-2,
+                                            optimizer='adam',
+                                            bias_optimizer='sgd',
+                                            weight_decay=1e-7,
+                                            loss='hinge')
+    elif request.param == 'mlp_mf_with_y_range':
+        model = MLPMatrixFactorizationModel(train=train,
+                                            val=val,
+                                            y_range=(0, 2))
     elif request.param == 'nonlinear_mf':
         model = NonlinearMatrixFactorizationModel(train=train,
                                                   val=val,
@@ -210,6 +253,33 @@ def models_trained_for_one_step(request,
         model = NeuralCollaborativeFiltering(train=train,
                                              val=val,
                                              final_layer=torch.tanh)
+    elif request.param == 'deep_fm':
+        model = DeepFM(train=train,
+                       val=val,
+                       embedding_dim=10,
+                       num_layers=1,
+                       dropout_p=0.1,
+                       lr=1e-3,
+                       weight_decay=0.,
+                       optimizer='adam',
+                       bias_optimizer=None,
+                       loss='hinge')
+    elif request.param == 'deep_fm_sigmoid':
+        model = DeepFM(train=train,
+                       val=val,
+                       final_layer='sigmoid')
+    elif request.param == 'deep_fm_relu':
+        model = DeepFM(train=train,
+                       val=val,
+                       final_layer='relu')
+    elif request.param == 'deep_fm_leaky_rulu':
+        model = DeepFM(train=train,
+                       val=val,
+                       final_layer='leaky_relu')
+    elif request.param == 'deep_fm_custom':
+        model = DeepFM(train=train,
+                       val=val,
+                       final_layer=torch.tanh)
     elif (
         request.param == 'hybrid_pretrained' or request.param == 'hybrid_pretrained_metadata_layers'
     ):
@@ -225,7 +295,7 @@ def models_trained_for_one_step(request,
                                                logger=False,
                                                checkpoint_callback=False)
         implicit_model_trainer.fit(implicit_model)
-        implicit_model.freeze()
+        implicit_model.eval()
 
         genres = (
             torch.tensor(movielens_metadata_df[
@@ -290,6 +360,6 @@ def models_trained_for_one_step(request,
     else:
         model_trainer.fit(model)
 
-    model.freeze()
+    model.eval()
 
     return model
